# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MDG
  version: '1'
  summary: Self-hosted personal accounting service.
  description: |
    MDG ("Moi.Den.Gi") is a self-hosted personal accounting web service. This
    document describes API v1 and is the only specification of that API — it is
    hand-written and hand-maintained, and nothing in the build verifies it
    against the code. If you change a controller, change this too.

    ## Content type

    Every request and response uses the media type
    `application/vnd.mdg+json;version=1` — errors included, and with no
    `charset` parameter. Write operations require it as a `Content-Type` request
    header, even the two maintenance triggers that take no body.

    The only responses without it are the `204 No Content` ones.

    ## Errors

    Errors are RFC 7807-shaped but carry the vendor media type rather than
    `application/problem+json`, and add a machine-readable `code` field.

    ## Authentication

    There is none. 
    
    ## Paging

    `GET /transactions` and `GET /budgets` are cursor-paged. Pass `limit` to
    start paging, then follow the `next` cursor. Cursors are **self-contained** —
    they carry the original filter, sort, embed and limit — so when `cursor` is
    supplied every other query parameter is ignored, though a `limit` below one
    is still rejected rather than quietly dropped. `first` and `next` come
    back as the empty string when there is no such page, and handing an empty
...

    ## Embedding

...

...
servers:
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List all transactions
      description: |
        Embedded accounts in this response always report `balance: 0` and
        `primary_balance: 0`; the balances are zeroed deliberately rather than
        computed per row.
      operationId: listTransactions
      parameters:
        - $ref: './components/parameters.yaml#/TransactionQuery'
        - $ref: './components/parameters.yaml#/TransactionSort'
        - $ref: './components/parameters.yaml#/EmbedAccount'
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: One page of transactions.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/transaction.yaml#/Transactions'
          links:
            GetListedTransaction:
              operationId: getTransaction
              description: Read the first transaction of this page.
              parameters:
                path.id: '$response.body#/transactions/0/id'
            UpdateListedTransaction:
              operationId: updateTransaction
              description: |
                Rewrite the first transaction of this page as a balanced pair of operations
                on the account it already touches.
              parameters:
                path.id: '$response.body#/transactions/0/id'
              requestBody:
                timestamp: '$response.body#/transactions/0/timestamp'
                comment: Updated through a link
                operations:
                  - account_id: '$response.body#/transactions/0/operations/0/account_id'
                    amount: -100
                    rate: 1
                  - account_id: '$response.body#/transactions/0/operations/0/account_id'
                    amount: 100
                    rate: 1
              x-schemathesis:
                merge_body: false
            DeleteListedTransaction:
              operationId: deleteTransaction
              description: Delete the first transaction of this page.
              parameters:
                path.id: '$response.body#/transactions/0/id'
            NextTransactionPage:
              operationId: listTransactions
              description: |
                Follow this page's `next` cursor. An exhausted page carries the empty
                string, which reads as *no cursor* and returns the first page again.
              parameters:
                query.cursor: '$response.body#/next'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
    post:
      tags: [Transactions]
      summary: Create a new transaction
      description: |
        A transaction is a balanced set of operations. The rules, each with the
        `code` it fails under:

        * Operations with an amount of exactly zero are **dropped** before
          validation, not rejected.
        * At least one non-zero operation must remain — `TRANSACTION_EMPTY`.
        * Every surviving operation must name an account. An `account_id` that is
          absent or `null` is rejected with 422 `TRANSACTION_DATA_INVALID`; one
          that names no account is 404 `ACCOUNT_NOT_FOUND`. A `null` in place of
          an operation object is the same 422.
        * `Σ(amount × rate)` over the remaining operations must be zero —
          `TRANSACTION_NOT_BALANCED`.
        * A rate may not be zero — `TRANSACTION_ZERO_RATE`.
        * In a multi-currency transaction **exactly one** currency acts as the
          conversion base and carries no rate. Every operation carrying a rate:
          `TRANSACTION_NO_DEFAULT_RATE`. No operation carrying one, or a mix of
          rate-less operations and operations with `rate: 1` across different
          currencies: `TRANSACTION_AMBIGUOUS_RATE`.

        For a single-currency transaction `rate` may be omitted throughout. The
        recommendation is to always send it, using `1` for the primary currency,
        so the base is explicit.

        Tags that do not exist yet are created.

      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/vnd.mdg+json;version=1:
            schema:
              $ref: './components/schemas/transaction.yaml#/Transaction'
            examples:
              singleCurrency:
                summary: Three rows, one currency, no rates needed
                value:
                  timestamp: '2017-02-04T16:45:36'
                  comment: Test transaction
                  tags: [test, transaction]
                  operations:
                    - { account_id: 1, amount: -150 }
                    - { account_id: 2, amount: 50 }
                    - { account_id: 3, amount: 100 }
              multiCurrency:
                summary: Two currencies — exactly one row omits `rate`
                value:
                  timestamp: '2017-02-05T13:54:35'
                  comment: Test transaction
                  tags: [test, transaction]
                  operations:
                    - { account_id: 2, amount: -100 }
                    - { account_id: 4, amount: 200, rate: 0.5 }
      responses:
        '201':
          description: The created transaction.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/transaction.yaml#/Transaction'
          links:
            GetTransaction:
              operationId: getTransaction
              description: Read the transaction that was just created.
              parameters:
                path.id: '$response.body#/id'
            UpdateTransaction:
              operationId: updateTransaction
              description: |
                Rewrite the transaction as a balanced pair of operations on the account it
                already touches, which re-runs every validation rule against a body that
                is known to pass them.
              parameters:
                path.id: '$response.body#/id'
              requestBody:
                timestamp: '$response.body#/timestamp'
                comment: Updated through a link
                operations:
                  - account_id: '$response.body#/operations/0/account_id'
                    amount: -100
                    rate: 1
                  - account_id: '$response.body#/operations/0/account_id'
                    amount: 100
                    rate: 1
              x-schemathesis:
                merge_body: false
            DeleteTransaction:
              operationId: deleteTransaction
              description: Delete the transaction that was just created.
              parameters:
                path.id: '$response.body#/id'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '404':
          description: No transaction found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
        '412':
          $ref: './components/responses.yaml#/PreconditionFailed'
        '415':
          $ref: './components/responses.yaml#/UnsupportedMediaType'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: View a transaction
      operationId: getTransaction
      parameters:
        - $ref: './components/parameters.yaml#/TransactionId'
        - $ref: './components/parameters.yaml#/EmbedAccount'
      responses:
        '200':
          description: The transaction.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/transaction.yaml#/Transaction'
        '404':
          description: No transaction found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
    put:
      tags: [Transactions]
      summary: Modify transaction data
      description: |
        Replaces the transaction wholesale, re-running every validation rule
        listed on `POST /transactions`. Account balances and any affected budget
        entries are recalculated.

      operationId: updateTransaction
      parameters:
        - $ref: './components/parameters.yaml#/TransactionId'
      requestBody:
        required: true
        content:
          application/vnd.mdg+json;version=1:
            schema:
              $ref: './components/schemas/transaction.yaml#/Transaction'
            example:
              timestamp: '2017-02-04T16:45:36'
              comment: Updated transaction
              tags: [test]
              operations:
                - { account_id: 1, amount: -80 }
                - { account_id: 3, amount: 80 }
      responses:
        '202':
          description: The updated transaction.
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/transaction.yaml#/Transaction'
        '400':
          $ref: './components/responses.yaml#/BadRequest'
        '404':
          description: No transaction found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
        '412':
          $ref: './components/responses.yaml#/PreconditionFailed'
        '415':
          $ref: './components/responses.yaml#/UnsupportedMediaType'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      description: |
        Tags left unused by the deletion are not removed from the tag dictionary.
      operationId: deleteTransaction
      parameters:
        - $ref: './components/parameters.yaml#/TransactionId'
      responses:
        '204':
          description: The transaction was deleted.
        '404':
          description: No transaction found
          content:
            application/vnd.mdg+json;version=1:
              schema:
                $ref: './components/schemas/problem.yaml#/Problem'
...